xnvmebuild: adjust to changes in xNVMe build-system
The xNVMe build-system has changed build-flags. Since the xNVMe fio
engine got upstream, then the external engine is removed, and with it,
the fio option to the build system. Additionally, then the
library-dependencies such as libvfn were boolean-options before, they
are now auto-detected features.

The xnvmebuild in SPDK is adjusted accordingly, removing the fio-flag,
and change the disablement of libvfn from 'false' to 'disabled'.

Signed-off-by: Karl Bonde Torp <k.torp@samsung.com>